Move peer and getters to CFLayer
