+++ src/java.desktop/macosx/classes/sun/java2d/opengl/CGLLayer.java
@@ -26,13 +26,10 @@
 package sun.java2d.opengl;
 
 import java.awt.GraphicsConfiguration;
-import java.awt.Rectangle;
-import java.awt.Transparency;
-
 import sun.awt.CGraphicsConfig;
 import sun.java2d.NullSurfaceData;
-import sun.java2d.SurfaceData;
 import sun.lwawt.LWWindowPeer;
+import sun.java2d.SurfaceData;
 import sun.lwawt.macosx.CFLayer;
 
 public class CGLLayer extends CFLayer {
@@ -42,7 +39,6 @@ public class CGLLayer extends CFLayer {
     private static native void validate(long layerPtr, CGLSurfaceData cglsd);
     private static native void blitTexture(long layerPtr);
 
-    private LWWindowPeer peer;
     private int scale = 1;
 
     public CGLLayer(LWWindowPeer peer) {
@@ -50,26 +46,6 @@ public class CGLLayer extends CFLayer {
 
         setPtr(nativeCreateLayer());
         this.peer = peer;
-    }
-
-    public Rectangle getBounds() {
-        return peer.getBounds();
-    }
-
-    public GraphicsConfiguration getGraphicsConfiguration() {
-        return peer.getGraphicsConfiguration();
-    }
-
-    public boolean isOpaque() {
-        return !peer.isTranslucent();
-    }
-
-    public int getTransparency() {
-        return isOpaque() ? Transparency.OPAQUE : Transparency.TRANSLUCENT;
-    }
-
-    public Object getDestination() {
-        return peer.getTarget();
     }
 
     public SurfaceData replaceSurfaceData() {
